requirements.txt
# requirements.txt

# Core LangChain and LangGraph
langchain
langchain-openai
langgraph
langchain-core
langchain-community

# Database
sqlalchemy[asyncio]
asyncpg

# FastAPI and web server
fastapi
uvicorn[standard]

# Utilities
python-dotenv
pydantic
colorama

# Optional: For better async support
aiofiles